Moved snap builds helper functions to separate module
